Backend/classifier/requirements.txt
django>4.0.0,<4.1.0
djangorestframework
django-cors-headers
django-filter
django-rest-auth
django-requests
tensorflow
djangorestframework
pillow
